requirements.txt
line-bot-sdk
flask
bs4
requests
imgurpython
psycopg2
gunicorn
